+++ src/rt/notify.rs
@@ -3,6 +3,8 @@ use crate::rt::{self, Access, Synchronize, VersionVec};
 
 use std::sync::atomic::Ordering::{Acquire, Release};
 
+use tracing::trace;
+
 #[derive(Debug, Copy, Clone)]
 pub(crate) struct Notify {
     state: object::Ref<State>,
@@ -16,7 +18,7 @@ pub(super) struct State {
     /// True if the notify woke up spuriously last time
     did_spur: bool,
 
-    /// When true, notification is sequentiall consistent.
+    /// When true, notification is sequential consistent.
     seq_cst: bool,
 
     /// `true` if there is a pending notification to consume.
@@ -40,6 +42,8 @@ impl Notify {
                 last_access: None,
                 synchronize: Synchronize::new(),
             });
+
+            trace!(?state, ?seq_cst, ?spurious, "Notify::new");
 
             Notify { state }
         })
@@ -70,6 +74,8 @@ impl Notify {
                     .map(|operation| operation.object());
 
                 if obj == Some(self.state.erase()) {
+                    trace!(state = ?self.state, thread = ?thread.id, "Notify::notify");
+
                     thread.unpark(active);
                 }
             }
@@ -90,6 +96,7 @@ impl Notify {
                 state.did_spur = true;
             }
 
+            trace!(state = ?self.state, notified = ?state.notified, ?spurious, "Notify::wait 1");
             dbg!((state.notified, spurious))
         });
 
@@ -107,6 +114,7 @@ impl Notify {
 
         // Thread was notified
         super::execution(|execution| {
+            trace!(state = ?self.state, "Notify::wait 2");
             let state = self.state.get_mut(&mut execution.objects);
 
             assert!(state.notified);