+++ src/sync/atomic/ptr.rs
@@ -8,7 +8,7 @@ pub struct AtomicPtr<T>(Atomic<*mut T>);
 
 impl<T> AtomicPtr<T> {
     /// Creates a new instance of `AtomicPtr`.
-    #[cfg_attr(loom_nightly, track_caller)]
+    #[track_caller]
     pub fn new(v: *mut T) -> AtomicPtr<T> {
         AtomicPtr(Atomic::new(v, location!()))
     }
@@ -19,37 +19,37 @@ impl<T> AtomicPtr<T> {
     }
 
     /// Get access to a mutable reference to the inner value.
-    #[cfg_attr(loom_nightly, track_caller)]
+    #[track_caller]
     pub fn with_mut<R>(&mut self, f: impl FnOnce(&mut *mut T) -> R) -> R {
         self.0.with_mut(f)
     }
 
     /// Loads a value from the pointer.
-    #[cfg_attr(loom_nightly, track_caller)]
+    #[track_caller]
     pub fn load(&self, order: Ordering) -> *mut T {
         self.0.load(order)
     }
 
     /// Stores a value into the pointer.
-    #[cfg_attr(loom_nightly, track_caller)]
+    #[track_caller]
     pub fn store(&self, val: *mut T, order: Ordering) {
         self.0.store(val, order)
     }
 
     /// Stores a value into the pointer, returning the previous value.
-    #[cfg_attr(loom_nightly, track_caller)]
+    #[track_caller]
     pub fn swap(&self, val: *mut T, order: Ordering) -> *mut T {
         self.0.swap(val, order)
     }
 
     /// Stores a value into the pointer if the current value is the same as the `current` value.
-    #[cfg_attr(loom_nightly, track_caller)]
+    #[track_caller]
     pub fn compare_and_swap(&self, current: *mut T, new: *mut T, order: Ordering) -> *mut T {
         self.0.compare_and_swap(current, new, order)
     }
 
     /// Stores a value into the pointer if the current value is the same as the `current` value.
-    #[cfg_attr(loom_nightly, track_caller)]
+    #[track_caller]
     pub fn compare_exchange(
         &self,
         current: *mut T,
@@ -61,7 +61,7 @@ impl<T> AtomicPtr<T> {
     }
 
     /// Stores a value into the atomic if the current value is the same as the current value.
-    #[cfg_attr(loom_nightly, track_caller)]
+    #[track_caller]
     pub fn compare_exchange_weak(
         &self,
         current: *mut T,
@@ -71,11 +71,33 @@ impl<T> AtomicPtr<T> {
     ) -> Result<*mut T, *mut T> {
         self.compare_exchange(current, new, success, failure)
     }
+
+    /// Fetches the value, and applies a function to it that returns an optional new value. Returns
+    /// a [`Result`] of [`Ok`]`(previous_value)` if the function returned [`Some`]`(_)`, else
+    /// [`Err`]`(previous_value)`.
+    #[track_caller]
+    pub fn fetch_update<F>(
+        &self,
+        set_order: Ordering,
+        fetch_order: Ordering,
+        f: F,
+    ) -> Result<*mut T, *mut T>
+    where
+        F: FnMut(*mut T) -> Option<*mut T>,
+    {
+        self.0.fetch_update(set_order, fetch_order, f)
+    }
 }
 
 impl<T> Default for AtomicPtr<T> {
     fn default() -> AtomicPtr<T> {
         use std::ptr;
         AtomicPtr::new(ptr::null_mut())
+    }
+}
+
+impl<T> From<*mut T> for AtomicPtr<T> {
+    fn from(p: *mut T) -> Self {
+        Self::new(p)
     }
 }