+++ tests/atomic_int.rs
@@ -24,6 +24,34 @@ macro_rules! test_int {
             }
 
             #[test]
+            fn max() {
+                loom::model(|| {
+                    let a: $int = NUM_A as $int;
+                    let b: $int = NUM_B as $int;
+
+                    let atomic = <$atomic>::new(a);
+                    let prev = atomic.fetch_max(b, SeqCst);
+
+                    assert_eq!(a, prev, "prev did not match");
+                    assert_eq!(a.max(b), atomic.load(SeqCst), "load failed");
+                });
+            }
+
+            #[test]
+            fn min() {
+                loom::model(|| {
+                    let a: $int = NUM_A as $int;
+                    let b: $int = NUM_B as $int;
+
+                    let atomic = <$atomic>::new(a);
+                    let prev = atomic.fetch_min(b, SeqCst);
+
+                    assert_eq!(a, prev, "prev did not match");
+                    assert_eq!(a.min(b), atomic.load(SeqCst), "load failed");
+                });
+            }
+
+            #[test]
             fn compare_exchange() {
                 loom::model(|| {
                     let a: $int = NUM_A as $int;
@@ -51,6 +79,19 @@ macro_rules! test_int {
                     assert_eq!(b, atomic.load(SeqCst));
                 });
             }
+
+            #[test]
+            fn fetch_update() {
+                loom::model(|| {
+                    let a: $int = NUM_A as $int;
+                    let b: $int = NUM_B as $int;
+
+                    let atomic = <$atomic>::new(a);
+                    assert_eq!(Ok(a), atomic.fetch_update(SeqCst, SeqCst, |_| Some(b)));
+                    assert_eq!(Err(b), atomic.fetch_update(SeqCst, SeqCst, |_| None));
+                    assert_eq!(b, atomic.load(SeqCst));
+                });
+            }
         }
     };
 }
@@ -59,3 +100,14 @@ test_int!(atomic_u8, u8, AtomicU8);
 test_int!(atomic_u16, u16, AtomicU16);
 test_int!(atomic_u32, u32, AtomicU32);
 test_int!(atomic_usize, usize, AtomicUsize);
+
+test_int!(atomic_i8, i8, AtomicI8);
+test_int!(atomic_i16, i16, AtomicI16);
+test_int!(atomic_i32, i32, AtomicI32);
+test_int!(atomic_isize, isize, AtomicIsize);
+
+#[cfg(target_pointer_width = "64")]
+test_int!(atomic_u64, u64, AtomicU64);
+
+#[cfg(target_pointer_width = "64")]
+test_int!(atomic_i64, i64, AtomicI64);